Move Data Types into dedicated data directory
