+++ src/engine/QueryExecutionTree.h
@@ -9,7 +9,7 @@
 #include <unordered_map>
 #include <unordered_set>
 
-#include "../parser/VarOrTerm.h"
+#include "../parser/data/VarOrTerm.h"
 #include "../util/Conversions.h"
 #include "../util/HashSet.h"
 #include "../util/streamable_generator.h"
